refactor: test organization
